Layout like a pip package
As a user, I would like to install gogeta-configure as a pip package, I
would like to run tests and I would like to be able to install in a
virtualenv.

This lays the foundation for installation via pip and running tests with
nose. I'll be adding the tests in a separate commit.